Mention that ING-4 supports GPL compliance
While reading, it occurred to me that ING-4 is necessary for compliance
with the GPL.

To paraphrase it, the GPL requires that sources for a GPL licensed
component be distributed along with the binaries built from those
sources or that include those sources. (I am not a lawyer.)

In order to comply with that redistribution requirement, you have to
capture and mirror those sources, which is what ING-4 is about.

Signed-off-by: Ralph Bean <rbean@redhat.com>